// packages/fragment/src/grammars/peggy-types.ts
export type Combinator = "+" | ">";
export type Property = IDENT;
export type Selector =
    | {
          type: "Selector";
          combinator: Combinator;
          left: SimpleSelector;
          right: Selector;
      }
    | {
          type: "Selector";
          combinator: " ";
          left: SimpleSelector;
          right: Selector;
      }
    | SimpleSelector;
export type SimpleSelector =
    | {
          type: "SimpleSelector";
          element: ElementName;
          qualifiers: (Id | Class | Attrib | Pseudo)[];
      }
    | {
          type: "SimpleSelector";
          element: "*";
          qualifiers: (Id | Class | Attrib | Pseudo)[];
      };
export type Id = { type: "IDSelector"; id: HASH };
export type Class = { type: "ClassSelector"; class: IDENT };
export type ElementName = IDENT | "*";
export type Attrib = {
    type: "AttributeSelector";
    attribute: IDENT;
    operator: string | null;
    value: string | null;
};
export type Pseudo = {
    type: "PseudoSelector";
    value: { type: "Function"; name: FUNCTION; params: IDENT[] } | IDENT;
};
// ...
export type IDENT = Ident;
export type HASH = string;
// ...
export type FUNCTION = Ident;